+++ lib/s3/s3-service.ts
@@ -1,4 +1,4 @@
-import { PutObjectCommand , S3Client } from "@aws-sdk/client-s3";
+import { PutObjectCommand, S3Client } from "@aws-sdk/client-s3";
 import { getSignedUrl } from "@aws-sdk/s3-request-presigner";
 
 // S3クライアントの初期化と共通のS3操作を管理するクラス
